repository initializer now works with batch ingest task

+++ persistence/src/main/java/edu/unc/lib/dl/services/DigitalObjectManagerImpl.java
@@ -880,8 +880,10 @@ public abstract class DigitalObjectManagerImpl implements DigitalObjectManager {
 		BatchIngestTask task = createBatchIngestTask(); // obtain from Spring prototype
 		task.init(prepDir);
 		task.run();
+		if(!task.isFailed()) {
+			FileUtils.deleteDir(prepDir);
+		}
 		task = null;
-		FileUtils.deleteDir(prepDir);
 	}
 
 	/**
